Add concurrent execution and progress tracking
This commit introduces concurrent execution of the 'run_engineer' function using asyncio and concurrent.futures. This allows the function to run in the background while the API continues to serve other requests.

A global dictionary 'operation_status' has been added to track the status of each operation. The status is updated at the start and end of the 'run_engineer' function.

The '/progress' endpoint has been updated to accept an 'app_name' parameter and return the status of the operation for that app. This provides a way to track the progress of each operation.

Minor changes include making the 'hello_world' function asynchronous and removing some commented out code.