Add app listing and deletion functionality
This commit introduces two new features to the FastAPI application. The first is a new endpoint at "/apps" that lists all the created apps. It does this by listing all the directories in the projects directory.

The second feature is an endpoint at "/delete/{app_name}" that deletes an app with the given name. It accomplishes this by deleting the directory associated with the app name. If the app does not exist, it raises an HTTPException with a 404 status code.