Add operation progress tracking to API
This commit introduces a new feature to the API that allows tracking the progress of an operation. A new global dictionary, operation_progress, has been added to store the progress of each operation. The progress is calculated as a percentage of the total steps completed in the run_engineer function.

The report_progress function has been updated to return not only the status of the operation but also the percentage of progress. This will provide more detailed feedback to the user about the status of their request.