Refactor FastAPI routes and add list_files endpoint
This commit includes a significant refactor of the FastAPI routes in the app_routes.py file. The changes include removing redundant comments and improving the code readability.

Additionally, a new endpoint, list_files, has been added. This endpoint lists all the files in a given app's directory and returns them. The current user is determined using the Depends function from FastAPI. If the app does not exist, a 404 HTTPException is raised with an appropriate error message.

The commit also includes minor changes to the run_prompt function, where the status of the operation is updated in the global dictionary.