Add docstrings to FastAPI routes
Added detailed docstrings to all the FastAPI routes in the app_routes.py file. These docstrings explain the purpose of each function, the parameters they take, and the type of response they return. This will make the code easier to understand and maintain. Also, corrected a minor typo in the list_files function docstring.