Add API key encryption and storage
This commit introduces the ability to store and retrieve encrypted API keys for users. The cryptography library is used to generate a Fernet encryption key, which is then used to encrypt the API keys before they are stored in a JSON file. Two new functions, get_api_key and set_api_key, have been added to handle the retrieval and storage of the keys. The cryptography library has also been added to the requirements.txt file.