Add option to improve code outside workspace
This commit introduces a new option "--improve-plus" that allows the user to improve code from an existing project that is located outside of the workspace folder. This is useful when the user wants to work on a codebase that is not in the default workspace. The workspace path is now set based on whether this new option is used or not.