Change README.md to LAST_MODIFICATION_README.md
The file name "README.md" has been changed to "LAST_MODIFICATION_README.md" in the parse_chat and overwrite_files functions. This change is reflected in the appending of the file name to the files list and the conditional check in the overwrite_files function. This is to ensure that the last modification readme file is correctly identified and updated. Also to allow to ask the LLM to update the README.md, if any in the project