Refactor AIAppsforge class and optimize threading
This commit refactors the AIAppsforge class by moving the initialization of the streaming_handler to the constructor. This change improves the code structure and makes the streaming_handler available immediately after an instance of AIAppsforge is created.

In addition, the commit optimizes the threading in the run_llm_and_get_tokens function. The sleep time has been reduced from 1 second to 0.05 seconds, which should make the function more responsive and efficient.

Unnecessary commented-out code and redundant sleep calls have also been removed to clean up the code.