Change return type of get_tokens method
The get_tokens method in the StreamInterceptor class was previously returning a List of strings. This has been changed to return a string instead. Also, in the run_llm_and_get_tokens function, the way tokens are added to output_tokens has been modified to use the append method instead of the += operator. This change aligns with the updated return type of the get_tokens method.