Comment out unused constant import
The import of MODEL_COST_PER_1K_TOKENS from langchain.callbacks.openai_info
is no longer needed in the extentions.py module. This change comments out
the unused import to clean up the code and potentially reduce confusion
about its usage. Further cleanup may be warranted if the import will not
be used in the future.